backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
sqlalchemy
python-dotenv
pydantic